Fix(Field): do not destroy dropdown if used by another container (#868)
* Fix(Field): do not destroy dropdown if used by another container

+++ inc/field.class.php
@@ -343,6 +343,11 @@ class PluginFieldsField extends CommonDBChild
     // phpcs:ignore PSR1.Methods.CamelCapsMethodName
     public function pre_deleteItem()
     {
+        /**
+         * @var \DBmysql $DB
+         */
+        global $DB;
+
         //retrieve search option ID to clean DiplayPreferences
         $container_obj = new PluginFieldsContainer();
         $container_obj->getFromDB($this->fields['plugin_fields_containers_id']);
@@ -389,7 +394,25 @@ class PluginFieldsField extends CommonDBChild
         ]);
 
         if ($this->fields['type'] === 'dropdown') {
-            return PluginFieldsDropdown::destroy($this->fields['name']);
+            //load all container (except current one) and check if another use this fields
+            $container_obj = new PluginFieldsContainer();
+            $all_container = $container_obj->find(['id' => ['!=', $this->fields['plugin_fields_containers_id']]]);
+
+            $use_by_another = false;
+            foreach ($all_container as $container_fields) {
+                foreach (json_decode($container_fields['itemtypes']) as $itemtype) {
+                    $dropdown_classname = PluginFieldsDropdown::getClassname($this->fields['name']);
+                    $classname = PluginFieldsContainer::getClassname($itemtype, $container_fields['name']);
+                    $dropdown_fk = getForeignKeyFieldForItemType($dropdown_classname);
+                    if ($DB->fieldExists(getTableForItemType($classname), $dropdown_fk)) {
+                        $use_by_another = true;
+                    }
+                }
+            }
+
+            if (!$use_by_another) {
+                return PluginFieldsDropdown::destroy($this->fields['name']);
+            }
         }
 
         return true;
